+++ BE/src/main/java/codesquad/issueTracker/oauth/dto/OauthJwt.java
@@ -1,0 +1,18 @@
+package codesquad.issueTracker.oauth.dto;
+
+public class OauthJwt {
+    public static String tokenType = "Bearer";
+    public String jwt;
+
+    public OauthJwt(String jwt) {
+        this.jwt = jwt;
+    }
+
+    public String getJwt() {
+        return jwt;
+    }
+
+    public String getTokenType() {
+        return tokenType;
+    }
+}
